changed file structure
added router and inbox page
