+++ VueADW/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "@tailwindcss/vite": "^4.1.3",
     "tailwindcss": "^4.1.3",
-    "vue": "^3.5.13"
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.1"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.2.3",
