+++ VueADW/src/pages/inbox.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+      <h1>Inbox</h1>
+      <p>This is the inbox page.</p>
+    </div>
+  </template>
+  
+  <script>
+  export default {
+    name: 'inbox'
+  }
+  </script>
+
