+++ VueADW/src/router/index.js
@@ -1,0 +1,23 @@
+import { createRouter, createWebHistory } from 'vue-router'
+import inbox from '../pages/inbox.vue'
+import home from '../pages/home.vue'
+
+const routes = [
+  {
+    path: '/',
+    name: 'home',
+    component: home
+  },
+  {
+    path: '/inbox',
+    name: 'inbox',
+    component: inbox
+  },
+]
+
+const router = createRouter({
+  history: createWebHistory(),
+  routes
+})
+
+export default router
